Load engine forms from disk
