--- src/server/forms/README.md
@@ -1,10 +1,0 @@
-# Pre-configured Forms
-
-This folder holds pre-configured form definitions that can be loaded by the runner:
-
-```js
-const server = await createServer({
-  formFileName: 'example.js',
-  formFilePath: join(cwd(), 'server/forms'),
-})
-```
