+++ src/server/plugins/engine/services/formsService.js
@@ -1,46 +1,43 @@
-import { formMetadataSchema } from '@defra/forms-model'
-
 import { config } from '~/src/config/index.js'
-import { FormStatus } from '~/src/server/routes/types.js'
-import { getJson } from '~/src/server/services/httpService.js'
+import { FileFormService } from '~/src/server/utils/file-form-service.js'
 
-/**
- * Retrieves a form definition from the form manager for a given slug
- * @param {string} slug - the slug of the form
- */
-export async function getFormMetadata(slug) {
-  const getJsonByType = /** @type {typeof getJson<FormMetadata>} */ (getJson)
-
-  const { payload: metadata } = await getJsonByType(
-    `${config.get('managerUrl')}/forms/slug/${slug}`
-  )
-
-  // Run it through the schema to coerce dates
-  const result = formMetadataSchema.validate(metadata)
-
-  if (result.error) {
-    throw result.error
-  }
-
-  return result.value
+// Create shared form metadata
+const now = new Date()
+const user = { id: 'user', displayName: 'Username' }
+const author = {
+  createdAt: now,
+  createdBy: user,
+  updatedAt: now,
+  updatedBy: user
+}
+const metadata = {
+  organisation: 'Defra',
+  teamName: 'Team name',
+  teamEmail: 'team@defra.gov.uk',
+  submissionGuidance: "Thanks for your submission, we'll be in touch",
+  notificationEmail: config.get('submissionEmailAddress'),
+  ...author,
+  live: author
 }
 
-/**
- * Retrieves a form definition from the form manager for a given id
- * @param {string} id - the id of the form
- * @param {FormStatus} state - the state of the form
- */
-export async function getFormDefinition(id, state) {
-  const getJsonByType = /** @type {typeof getJson<FormDefinition>} */ (getJson)
+// Instantiate the file loader form service
+const loader = new FileFormService()
 
-  const suffix = state === FormStatus.Draft ? `/${state}` : ''
-  const { payload: definition } = await getJsonByType(
-    `${config.get('managerUrl')}/forms/${id}/definition${suffix}`
-  )
+// Add a Json form
+await loader.addForm('src/server/forms/register-as-a-unicorn-breeder.json', {
+  ...metadata,
+  id: '95e92559-968d-44ae-8666-2b1ad3dffd31',
+  title: 'Register as a unicorn breeder',
+  slug: 'register-as-a-unicorn-breeder'
+})
 
-  return definition
-}
+// Add a Yaml form
+await loader.addForm('src/server/forms/register-as-a-unicorn-breeder.yaml', {
+  ...metadata,
+  id: '641aeafd-13dd-40fa-9186-001703800efb',
+  title: 'Register as a unicorn breeder (yaml)',
+  slug: 'register-as-a-unicorn-breeder-yaml' // if we needed to validate any JSON logic, make it available for convenience
+})
 
-/**
- * @import { FormDefinition, FormMetadata } from '@defra/forms-model'
- */
+// Get the forms service
+export const formsService = loader.toFormsService()
